galatea: Cleanup and modernize.
